wip.  create src/api and move tx_initiator into it.  so we can have more api modules
