+++ src/api/tx_initiation/mod.rs
@@ -68,7 +68,6 @@
 // these represent the public tx_initiator API
 pub mod builder;
 pub mod error;
-pub mod export;
 pub mod initiator;
 pub mod send;
