+++ src/bin/neptune-cli.rs
@@ -35,7 +35,7 @@ use neptune_cash::models::state::wallet::wallet_status::WalletStatusExportFormat
 use neptune_cash::rpc_auth;
 use neptune_cash::rpc_server::error::RpcError;
 use neptune_cash::rpc_server::RPCClient;
-use neptune_cash::tx_initiation::builder::tx_output_list_builder::OutputFormat;
+use neptune_cash::api::tx_initiation::builder::tx_output_list_builder::OutputFormat;
 use rand::Rng;
 use regex::Regex;
 use serde::Deserialize;
